+++ src/spinmdkit/analysis/__init__.py
@@ -2,14 +2,17 @@
 
 from .moment_series import MomentSeries, analyze_moment_series
 from .observables import magnetization, moment_norms, neel_vector, torques
+from .spatial import LayerSelection, select_layer
 from .summary import summarize_frame
 
 __all__ = [
+    "LayerSelection",
     "MomentSeries",
     "analyze_moment_series",
     "magnetization",
     "moment_norms",
     "neel_vector",
+    "select_layer",
     "summarize_frame",
     "torques",
 ]